Move multivalue.h from searchlib to searchcommon.
